+++ course/src/main/java/org/openedx/course/presentation/ui/CourseUI.kt
@@ -39,6 +39,7 @@ import androidx.compose.material.LinearProgressIndicator
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.OutlinedButton
 import androidx.compose.material.Scaffold
+import androidx.compose.material.Snackbar
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
@@ -1153,6 +1154,59 @@ fun CourseDatesBannerTablet(
                 text = stringResource(id = it),
                 onClick = resetDates,
             )
+        }
+    }
+}
+
+@Composable
+fun DatesShiftedSnackBar(
+    showAction: Boolean = false,
+    onViewDates: () -> Unit? = {},
+    onClose: () -> Unit? = {},
+) {
+    Snackbar(
+        modifier = Modifier.padding(16.dp),
+        backgroundColor = MaterialTheme.appColors.background
+    ) {
+        Column(modifier = Modifier.padding(4.dp)) {
+            Box {
+                Text(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .align(Alignment.CenterStart),
+                    text = stringResource(id = org.openedx.core.R.string.core_dates_shift_dates_successfully_title),
+                    color = MaterialTheme.appColors.textFieldText,
+                    style = MaterialTheme.appTypography.titleMedium
+                )
+                IconButton(modifier = Modifier.align(Alignment.TopEnd), onClick = { onClose() }) {
+                    Icon(
+                        imageVector = Icons.Filled.Close,
+                        contentDescription = "close",
+                        tint = MaterialTheme.appColors.onBackground,
+                    )
+                }
+            }
+            Text(
+                modifier = Modifier
+                    .padding(top = 4.dp)
+                    .fillMaxWidth(),
+                text = stringResource(id = org.openedx.core.R.string.core_dates_shift_dates_successfully_msg),
+                color = MaterialTheme.appColors.textFieldText,
+                style = MaterialTheme.appTypography.titleSmall,
+            )
+            if (showAction) {
+                OpenEdXOutlinedButton(
+                    modifier = Modifier
+                        .padding(top = 16.dp)
+                        .fillMaxWidth(),
+                    text = stringResource(id = org.openedx.core.R.string.core_dates_view_all_dates),
+                    backgroundColor = MaterialTheme.appColors.background,
+                    textColor = MaterialTheme.appColors.primary,
+                    borderColor = MaterialTheme.appColors.primary,
+                    onClick = {
+                        onViewDates()
+                    })
+            }
         }
     }
 }
