+++ course/src/main/res/values/strings.xml
@@ -51,6 +51,7 @@
     <string name="course_you_are_already_enrolled" tools:ignore="MissingTranslation">You are already enrolled in this course.</string>
     <string name="course_discover" tools:ignore="MissingTranslation">Discover</string>
     <string name="course_change_quality_when_downloading">You cannot change the download video quality when all videos are downloading</string>
+    <string name="course_dates_shifted_message">Dates Shifted</string>
 
     <!-- Error Message when course dates are not available -->
     <string name="course_dates_unavailable_message" tools:ignore="MissingTranslation">Course dates are not currently available.</string>
